requirements.txt
numpy
pandas
torch
scikit-learn
joblib
mpi4py
pygame
PyOpenGL
